Refactor File Names, leaves useful comments when navigating the project and its files
