+++ screens/TestScreen.js
@@ -21,24 +21,3 @@ export default function authScreen() {
   </View>
   );
 };
-
-
-
-
-// class SignInScreen extends React.Component {
-//   static navigationOptions = {
-//     title: 'Please sign in',
-//   };
-
-//   render() {
-//     return (
-//       <View style={styles.container}>
-//         <Button title="Sign in!" onPress={this._signInAsync} />
-//       </View>
-//     );
-//   }
-
-//   _signInAsync = async () => {
-//     await AsyncStorage.setItem('userToken', 'abc');
-//     this.props.navigation.navigate('App');
-//   };
